# todo_rest_specs/todo_swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the contract for our ToDo app. Data will be stored in Domino. The UI will be React",
    "version": "1.0.0",
    "title": "Swagger Domino/React ToDos",
    "contact": {
      "name": "Paul Withers",
      "email": "watsonwork.me/pwithers@intec.co.uk",
      "url": "http://www.intec.co.uk"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8081",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/todos/{store}/findByStatus": {
      "get": {
        "tags": [
          "todo"
        ],
        "description": "get all todos in a store for a particular status, sorted by due date ascending. If no status is passed, default will be incomplete",
        "operationId": "getToDosByStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "description": "store id or name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "description": "status",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "new",
              "incomplete",
              "complete",
              "reassigned"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "todos for that status",
            "schema": {
              "type": "array",
              "description": "array of todos",
              "items": {
                "$ref": "#/definitions/ToDo"
              }
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
    "/todos/{store}/findByAsigneeAndStatus": {
      "get": {
        "tags": [
          "todo"
        ],
        "description": "get all todos in a store assigned to a particular user for a particular status, sorted by due date ascending. If no username or status is passed, current user and incomplete status will be used",
        "operationId": "getToDosByAssigneeAndStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "description": "store id or name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "description": "username",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "description": "status",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "new",
              "incomplete",
              "complete",
              "reassigned"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "todos assigned to the user for the status",
            "schema": {
              "type": "array",
              "description": "array of todos",
              "items": {
                "$ref": "#/definitions/ToDo"
              }
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
    "/todos/{store}/findByPriorityAndStatus": {
      "get": {
        "tags": [
          "todo"
        ],
        "description": "get all todos in a store for a particular priority and status, sorted by due date ascending. If no priority and status are passed, high and unassigned will be used",
        "operationId": "getToDosByPriorityAndStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "description": "store id or name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "priority",
            "description": "priority",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ]
          },
          {
            "name": "status",
            "description": "status",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "new",
              "incomplete",
              "complete",
              "reassigned"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "todos for the priority and status",
            "schema": {
              "type": "array",
              "description": "array of todos",
              "items": {
                "$ref": "#/definitions/ToDo"
              }
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
    "/todos/{store}/findByDate": {
      "get": {
        "tags": [
          "todo"
        ],
        "description": "get all todos in a store after or between dates",
        "operationId": "getToDosByDate",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "description": "store id or name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "startDate",
            "description": "start date in RFC 3339 section 5.6 format 2018-03-19",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "date"
          },
          {
            "name": "endDate",
            "description": "start date in RFC 3339 section 5.6 format 2018-03-19",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "incomplete todos for the priority",
            "schema": {
              "type": "array",
              "description": "array of todos",
              "items": {
                "$ref": "#/definitions/ToDo"
              }
            }
          },
          "400": {
            "description": "Invalid store name or id supplied"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "domino_auth": []
          },
          {
            "application_auth": [],
            "user_auth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "domino_auth": {
      "type": "basic"
    },
    "application_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-TODO-API-KEY"
    },
    "user_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-TODO-USER-KEY"
    },
    "debug_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-DEBUG-KEY"
    },
    "sched_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-SCHED-KEY"
    }
  },
  "definitions": {
...
    "ToDo": {
      "type": "object",
      "properties": {
        "unid": {
          "type": "string",
          "description": "MetaversalID to access this ToDo from just a Domino Session. First 16 chars are replicaId, rest is UNID",
          "minLength": 48,
          "maxLength": 48,
          "format": "hex",
          "readOnly": true
        },
        "author": {
          "type": "string",
          "description": "username of the person who creates the ToDo - set automatically via the API from the currentUser passed across",
          "readOnly": true
        },
        "currentUser": {
          "$ref": "#/definitions/User"
        },
        "taskName": {
          "type": "string",
          "description": "Brief description of task"
        },
        "description": {
          "type": "string",
          "description": "Detailed description of the task"
        },
        "dueDate": {
          "type": "string",
          "format": "date",
          "description": "Date when the task is due in RFC 3339 section 5.6 format 2018-03-19"
        },
        "priority": {
          "type": "string",
          "description": "Priority for the ToDo, see enum",
          "enum": [
            "Low",
            "Medium",
            "High"
          ]
        },
        "assignedTo": {
          "type": "string",
          "description": "Name of the person the ToDo is assigned to. If not set, current user will be used"
        },
        "status": {
          "type": "string",
          "description": "Current status of the ToDo. Set via workflow actions, defaulting to New",
          "enum": [
            "New",
            "Reassigned",
            "Complete",
            "Overdue"
          ],
          "readOnly": true
        }
      },
      "required": [
        "currentUser",
        "taskName",
        "description",
        "dueDate",
        "priority"
      ]
    },
    "User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "description": "If Domino authentication is in operation for REST access, this property will be ignored for the CurrentUser. If not, the REST service  will use the username passed, replacing '@' with '_', append a computed OU, and use the server's O."
        },
        "access": {
          "type": "object",
          "description": "Required when setting access levels, returned when checking access levels, irrelevant otherwise.",
          "properties": {
            "level": {
              "type": "string",
              "description": "Application access level. This will be converted to Editor or Reader in Domino ACL",
              "enum": [
                "Admin",
                "Reader",
                "Editor",
                "No Access"
              ]
            },
            "allowDelete": {
              "type": "boolean",
              "description": "Whether or not Delete documents privilege is added"
            }
          }
        }
      }
    }
  }
}